tssupervisorupdate: Add compatible_id to support multi-board images
The i.MX93s can all use the same image. For example, the 0x9370's
Wizard images are compatible with the 0x9390.

An image's `modelnum` contains a single model ID. It probably should
be called `compatible_id`. The board struct, at least, will add that
field. Now, if the image doesn't match the board's model number, the
image will be checked against what the board's images are compatible
with.

In this way, an image will work if labeled with the board's `modelnum`
or what its (optional) `compatible_id` says it is compatible with.

All board's that can use the same image will have a `compatible_id`
field added. That includes the "parent" of the family, so it is clear
that its images are meant to be compatible with other boards.

If no `compatible_id` field exists, it is effectively ignored. As long
as an image doesn't have a "`modelnum`" that's zero!